requirements.txt
fastapi
fastapi-cli
uvicorn
motor
pydantic
pydantic[email]  
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pika
fpdf==1.7.2
pre-commit == 3.7.0
black==24.3.0
flake8==6.1.0
flake8-django
isort==5.13.2
psycopg2-binary
alembic
sqlalchemy
psycopg2-binary
